When the COOKIEPATH and SITECOOKIEPATH in WordPress differ, but are in the same path hierarchy, Wordpress doesn't care and still creates a wordpress_logged_in cookie for each path.
In bbPress we just created one, which works, except it wouldn't clear the second WordPress generated one on logout, which is bad.

This change removes the check for these paths being in the same path hierarchy and just creates both (and also destroys both).

Fixes lots, but not actually ticketed.