Use pre_get_posts to filter out hidden forum content instead of using multiple calls to bbp_exclude_forum_ids().
This ensures that no hidden or private forum content leaks out into widgets, profiles, and views, and also limits the liability of future plugins from needing to exclude those forum_id's manually.

The behavior of bbp_exclude_forum_ids() has changed to handle the gathering of forum ID's to be excluded, with the optional parameter to return results in different formats.