Users on single-site installations can have all-numeric usernames, causing profile URLs like '/forums/users/55555/' to be parsed incorrectly.
In bbp_parse_query(), check the 'permalink_structure' option before checking if the 'bbp_user' WP query variable is not numeric.

This has the potentially negative side-effect of not being able to use the query-var with a numeric user ID when using pretty permalinks, but that was never intended behavior anyways.

Fixes #2136.