Remove `languages` folder, and move `bbpress.pot` back into the root plugin directory.
The existence of the `languages` folder only currently sets up a confusing experience for a growing subset of users that also cannot easily get support in their language. By deleting the directory, it's less likely users will put their translations in a folder that does not already exist, which prevents them being deleted on bbPress plugin updates.

Because we have no plans of packaging translations directly with bbPress itself, and because wp.org will manage the downloading and installation of translations for us (and all plugins) based on the language of the WordPress installation, deleting the `languages` folder will provide a better experience for our non-English users going forward.