Forums: Use `no_found_rows` by default in `bbp_has_forums()`
Forums do not currently support pagination, so the total number of rows is not traditionally useful for this loop. In cases where it is, it can be explicitly overloaded in the `$args` array.

This results in 1 less database query per page where a forums list is shown, and a speed improvement overall.