Moderators: Introduce "Super Moderators" setting.
This change introduces the option (and default off value), admin setting and UI for toggling it on and off.

A "Super Moderator" is a moderator with the ability to also edit users, a commonly requested and popular feature in other forum software. The use-case is that user accounts may be spammers and need to be moderated beyond just their topics & replies, but the `edit_users` capability is reserved for Administrators (or Super Admins in multisite.) This setting will act as the global on/off to the deeper implementation of this feature.