Forums: Use `bbp_get_forum_subforum_count()` in `bbp_forum_get_subforums()`.
This performance optimization will result in 1 less query per forum in the forums list that does not have subforums, by checking the subforum meta-data before trying to query for subforums that do not exist anyways.

This is less of an issue with a persistent object cache, but those without that luxury can now save those database queries for when they really matter.