Template: adjust escaping of pagination count strings.
This change fixes a regression (in trunk only) that caused the filtered results of pagination counts to always be late escaped, but the intention was really only to escape the output of `_n()` which does not have an escaped equivalent function in the Gettext API.

I also tweaked the logic in `bbp_get_topic_pagination_count()` to not bail early. This brings it inline with `bbp_get_forum_pagination_count()` and allows the filter to run even when the text string is empty. Flexibility is the goal, so skipping the filter was also unintentional, and is now fixed.

Fixes #3162.