Admin: add an admin notice if database upgrade was skipped.
This change adds a persistent admin notice (who doesn't love those?) in the event that the forums are part of a large WordPress installation that prevented the automatic database upgrade routine from running.

Two links are provided: one to "Go Upgrade" and another to "Hide Forever". The first will take the user (with adequate capability) to the `Tools > Forums > Upgrade` screen; the second will delete the notice nag key from `wp_options`.

This includes some admin area CSS tweaks, so the asset version gets a bump as well.