Common: remove redundant calls to `bbp_number_not_negative()`.
This change decreases the filter priority of out-of-range prevention from `10` to `8`, and also adds filters to pre-formatted values in lieu of hard-coding this function into each individual count getter. This allows plugins to unplug this check more easily, and makes sure each value is only checked once for each call instead of occasionally twice in a row.