Anonymous: remove escaped placeholders from `$where` clause of duplicate check.
This change fixes a bug for versions of WordPress 4.8.3 and higher, that was causing the anonymous email comparison to use the placeholder text instead of the raw email address, but only when an email address contained a `%` character.

Also use the correct return variant, and make sure the output is escaped.