Reset: DRY out the `bbp_admin_reset_database()` logic.
This change introduces a helper function for running queries and producing user feedback based on those results. It also introduces a `Skipped` result if the query is invalid or empty. Ultimately, this results in fewer repeated lines of code, fewer reused local variables, and an easier to understand code flow.

An old, leftover routine for removing `forum-mod` taxonomy terms was also removed, as that approach was abandoned during the 2.6 development cycle but was missed when other refactoring was done.