Subscriptions: ensure special characters in author names are encoded properly.
This commit copies the approach used on the content, and applies `wp_specialchars_decode()` and `strip_tags()` to author names in subscription emails.

(HTML tags are not usually allowed in usernames, but we strip them anyways in the event a filter allowed them. Better to keep emails legible than include unintended tags?)

Fixes #3256.