Upgrades: Link to pending upgrades, if there are any.
This commit modifies the "Upgrade Forums" page by adding Statuses to filter by, specifically for pending upgrades. This allows bbPress to link directly to the upgrades that need doing, rather than the generic page itself.

Along with this commit are accompanying styling tweaks and helper functions to make this functionality possible.

Fixes #3244.