Administration: allow post types to support `custom-fields`.
This commit fixes a regression in 2.6.x by removing the `#the-list` identifier from the Replies list-table inside of Topic Edit meta-box (phew!) which was causing a duplicate ID collision with the Custom Fields list-table, resulting in the JavaScript for managing them to no longer function.

It subsequently removes all references to `#the-list` in CSS to completely avoid using this ID for anything, and uses `tbody` instead for the same effect but with reduced priority (which is a good thing here).

In trunk for 2.7.0. See #3311.