Topic Views: do not sanitize `$view` parameter of `bbp_get_view_id()`.
(Removes the call to `sanitize_key()` inside of `bbp_get_view_id()`.)

This change ensures that when checking for a registered view, the literal value is used for comparisons and not the sanitized version of it. This matches the behavior of other `_get_*_id()` functions, and also fixes a bug causing these invalid view IDs not to 404 as intended.

Props johnjamesjacoby, dd32.

In trunk, for 2.7.

Fixes #3438.