Posts: strip "Protected/Private" for bbPress post types
This change prevents post title hints from being prepended to titles of Forums, Topics, or Replies, which was happening as a consequence of calling `get_the_title()` which assumes every protected/private post title needs it.

It does this by including the following code changes:

* introduce the `bbp_no_title_status_hints()` filter function, which hooks into 2 WordPress filters and maybe overrides the return value
* replace a few `get_the_title()` calls with `get_post_field( 'post_title' )` to bypass the above filters when they would never be desirable

Committer note: this could be considered a small back-compat break (because it alters long-standing theme-side output) but ultimately it is a design decision to output all of the forum content as it was saved by the community members. Forums that prefer the old behavior can unhook `bbp_no_title_status_hints` and continue to customize child template parts to insert custom forum status hints.

Fixes #3602.