Breadcrumbs: prevent duplicate root crumb on Forum Archive subpages
This change prevents the Forum Archive page from appearing a second time in the breadcrumb trail when using pages, shortcodes, and/or subpages for forum setup.

It does this by keeping track of a previous match of a page by path to the Forum Archive, and skipping that page ID in the subsequent ancestors loop.

In trunk, for 2.7.