Common: prefer `bbp_get_topic_permalink()` over `get_permalink()`
This change ensures that features like Subscriptions & Pagination (inside of topics list loops) have: an opportunity to use the anticipated function path, and access to the right filters when including links back to them.

This fixes a potential bug where filtering a topic permalink would not happen inside of emails or pagination in the loop-topic template part.

See #3603.