Core: small updates to bbPress()->data
This change tweaks the code commentary a bit, and adds a default value of an empty array.

This is largely for the sake of code cleanliness, but also removes the potential for debug notices or errors inside of the magic methods from `data` not being initialized as an array when they are first called.